detach: prompt for user confirmation before detach

+++ uaclient/cli.py
@@ -242,6 +242,8 @@ def action_detach(args, cfg):
         print("Detach will disable the following service{}:".format(suffix))
         for ent in to_disable:
             print("    {}".format(ent.name))
+    if not util.prompt_for_confirmation():
+        return 0
     for ent in to_disable:
         ent.disable(silent=True)
     cfg.delete_cache()
